refactor: move refresh url from backend to batch
This is the first step toward removing sessions. Which needlessly complicate the implementation. More on that in subsequent commit messages.
